+++ R/draw.R
@@ -51,6 +51,8 @@ drawDependencies <- function(drag = TRUE) {
 #'
 #' ## for more examples see
 #' # browseURL(system.file("examples/draw.R", package = "leaflet.extras"))
+#' # browseURL(system.file("examples/shiny/draw-events/app.R", package = "leaflet.extras"))
+#' # browseURL(system.file("examples/shiny/draw-events/draw_mouse_events.R", package = "leaflet.extras"))
 addDrawToolbar <- function(
     map, targetLayerId = NULL, targetGroup = NULL,
     position = c("topleft", "topright", "bottomleft", "bottomright"),